LangChain_system/requirements.txt
openai==1.45.0
faiss-cpu==1.12.0
pypdf2==3.0.1
python-dotenv==1.0.0
sentence-transformers==5.1.1